requirements.txt
pymongo
selenium
python-dotenv